+++ cli/src/commands/status.rs
@@ -52,6 +52,9 @@ pub(crate) struct StatusArgs {
     /// Restrict the status display to these paths
     #[arg(value_name = "FILESETS", value_hint = clap::ValueHint::AnyPath)]
     paths: Vec<String>,
+    /// Show ignored files as well.
+    #[arg(short, long, action = clap::ArgAction::SetTrue)]
+    ignored: bool,
 }
 
 #[instrument(skip_all)]
@@ -85,6 +88,7 @@ pub(crate) async fn cmd_status(
 
         let wc_has_changes = tree.tree_ids() != parent_tree.tree_ids();
         let wc_has_untracked = !snapshot_stats.untracked_paths.is_empty();
+        let wc_has_ignored = !snapshot_stats.ignored_paths.is_empty();
         if !wc_has_changes && !wc_has_untracked {
             writeln!(formatter, "The working copy has no changes.")?;
         } else {
@@ -131,6 +135,29 @@ pub(crate) async fn cmd_status(
                 )
                 .await?;
             }
+        }
+
+        if args.ignored && wc_has_ignored {
+            writeln!(formatter, "Ignored paths:")?;
+            visit_collapsed_untracked_files(
+                snapshot_stats.ignored_paths,
+                snapshot_stats.untracked_paths.keys(),
+                &tree,
+                |path, is_dir| {
+                    let ui_path = workspace_command.path_converter().format_file_path(path);
+                    writeln!(
+                        formatter.labeled("diff").labeled("ignored"),
+                        "! {ui_path}{}",
+                        if is_dir {
+                            std::path::MAIN_SEPARATOR_STR
+                        } else {
+                            ""
+                        }
+                    )?;
+                    Ok(())
+                },
+            )
+            .await?;
         }
 
         let template = workspace_command.commit_summary_template();